Split DataFlowPass.cs into two classes
