+++ src/Compilers/CSharp/Portable/FlowAnalysis/DataFlowsOutWalker.cs
@@ -183,7 +183,7 @@ namespace Microsoft.CodeAnalysis.CSharp
         }
 #endif
 
-        protected override void AssignImpl(BoundNode node, BoundExpression value, bool? valueIsNotNull, RefKind refKind, bool written, bool read)
+        protected override void AssignImpl(BoundNode node, BoundExpression value, RefKind refKind, bool written, bool read)
         {
             if (IsInside)
             {
@@ -209,7 +209,7 @@ namespace Microsoft.CodeAnalysis.CSharp
                 }
             }
 
-            base.AssignImpl(node, value, valueIsNotNull, refKind, written, read);
+            base.AssignImpl(node, value, refKind, written, read);
         }
 
         private bool FlowsOut(ParameterSymbol param)
